Batch update setup scripts (#2541)

+++ tools/setup_nvidia.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 
-# Override triton
-uv pip uninstall triton
-uv pip install --index ${FLAGOS_PYPI} \
-    flagtree==0.5.0+3.5
+uv pip install -e . flag_gems[nvidia,test]
 
-uv pip install -e .[nvidia,test]
+if [ -n "${USE_FLAGTREE}" ]; then
+  uv pip uninstall triton
+  uv pip install --index ${FLAGOS_PYPI} \
+    flagtree==0.5.0+3.5
+fi
